+++ packages/webcodecs/src/on-audio-track.ts
@@ -6,9 +6,9 @@ import {getAudioEncoderConfig} from './audio-encoder-config';
 import type {ConvertMediaAudioCodec} from './codec-id';
 import {convertEncodedChunk} from './convert-encoded-chunk';
 import type {ConvertMediaContainer, ConvertMediaState} from './convert-media';
+import {defaultOnAudioTrackHandler} from './default-audio-track-handler';
 import Error from './error-cause';
 import type {ConvertMediaOnAudioTrackHandler} from './on-audio-track-handler';
-import {defaultOnAudioTrackHandler} from './on-audio-track-handler';
 
 export const makeAudioTrackHandler =
 	({
